[fix] 퀴즈 풀기 사이드 정보 섹션 컴포넌트 분리
